requirements.txt
# ============================================================================
# MusicElo Dependencies
# ============================================================================

# Core Framework
streamlit>=1.30.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0  # Database migrations (optional)

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Clients
spotipy>=2.23.0  # Spotify API
ytmusicapi>=1.3.0  # YouTube Music API

# Environment & Configuration
python-dotenv>=1.0.0

# Data Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0  # Progress bars for scripts

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality (Optional)
black>=23.0.0  # Code formatting
flake8>=6.1.0  # Linting
mypy>=1.5.0  # Type checking

# Development (Optional)
ipython>=8.12.0  # Better Python shell
jupyter>=1.0.0  # Notebooks for analysis
